Rework code for class libraries
- rename code files to shorten path
- rework CMakes accordingly
- rework native implementation to remove dependency of Interop
- update nanoCLR test solution to newest Nuget packages
