+++ CMake/Modules/FindWindows.Devices.Gpio.cmake
@@ -18,11 +18,9 @@ list(APPEND Windows.Devices.Gpio_INCLUDE_DIRS "${PROJECT_SOURCE_DIR}/${BASE_PATH
 # source files
 set(Windows.Devices.Gpio_SRCS
 
-    windows_devices_gpio_native_Windows_Devices_Gpio_GpioController.cpp
-    windows_devices_gpio_native_Windows_Devices_Gpio_GpioController_mshl.cpp
-    windows_devices_gpio_native_Windows_Devices_Gpio_GpioPin.cpp
-    windows_devices_gpio_native_Windows_Devices_Gpio_GpioPin_mshl.cpp
-    windows_devices_gpio_native.cpp
+    win_dev_gpio_native_Windows_Devices_Gpio_GpioController.cpp
+    win_dev_gpio_native_Windows_Devices_Gpio_GpioPin.cpp
+    win_dev_gpio_native.cpp
 )
 
 foreach(SRC_FILE ${Windows.Devices.Gpio_SRCS})
