--- targets/CMSIS-OS/ChibiOS/nanoCLR/Windows.Devices.Gpio/windows_devices_gpio_native_Windows_Devices_Gpio_GpioController_mshl.cpp
@@ -1,47 +1,0 @@
-//
-// Copyright (c) 2017 The nanoFramework project contributors
-// Portions Copyright (c) Microsoft Corporation.  All rights reserved.
-// See LICENSE file in the project root for full license information.
-//
-
-
-#include "windows_devices_gpio_native.h"
-#include "windows_devices_gpio_native_Windows_Devices_Gpio_GpioController.h"
-
-using namespace Windows::Devices::Gpio;
-
-
-HRESULT Library_windows_devices_gpio_native_Windows_Devices_Gpio_GpioController::NativeOpenpin___BOOLEAN__I4( CLR_RT_StackFrame& stack )
-{
-    NANOCLR_HEADER(); hr = S_OK;
-    {
-        CLR_RT_HeapBlock* pMngObj = Interop_Marshal_RetrieveManagedObject( stack );
-
-        FAULT_ON_NULL(pMngObj);
-
-        signed int param0;
-        NANOCLR_CHECK_HRESULT( Interop_Marshal_INT32( stack, 1, param0 ) );
-
-        bool retVal = GpioController::NativeOpenpin( pMngObj,  param0, hr );
-        NANOCLR_CHECK_HRESULT( hr );
-        SetResult_bool( stack, retVal );
-
-    }
-    NANOCLR_NOCLEANUP();
-}
-
-HRESULT Library_windows_devices_gpio_native_Windows_Devices_Gpio_GpioController::get_PinCount___I4( CLR_RT_StackFrame& stack )
-{
-    NANOCLR_HEADER(); hr = S_OK;
-    {
-        CLR_RT_HeapBlock* pMngObj = Interop_Marshal_RetrieveManagedObject( stack );
-
-        FAULT_ON_NULL(pMngObj);
-
-        signed int retVal = GpioController::get_PinCount( pMngObj,  hr );
-        NANOCLR_CHECK_HRESULT( hr );
-        SetResult_INT32( stack, retVal );
-
-    }
-    NANOCLR_NOCLEANUP();
-}
