--- targets/CMSIS-OS/ChibiOS/nanoCLR/nanoFramework.Runtime.Events/nanoframework_runtime_events_nanoFramework_Runtime_Events_EventSink_mshl.cpp
@@ -1,8 +1,0 @@
-//
-// Copyright (c) 2017 The nanoFramework project contributors
-// Portions Copyright (c) Microsoft Corporation.  All rights reserved.
-// See LICENSE file in the project root for full license information.
-//
-
-// this file is empty on purpose
-// because the specifics of the event sink the marshaling code of this library doesn't require implementation
