+++ targets/os/win32/nanoCLR/CLRStartup.cpp
@@ -147,14 +147,21 @@ struct Settings
 		// just need to update the path on the package folder as the version changes //
 		// ************************************************************************* //
 		vec.push_back(L"-load");
-		vec.push_back(L"..\\packages\\nanoFramework.Windows.Devices.Gpio.1.0.0-preview006\\lib\\Windows.Devices.Gpio.pe");
+		vec.push_back(L"..\\packages\\nanoFramework.Windows.Devices.Gpio.1.0.0-preview010\\lib\\Windows.Devices.Gpio.pe");
+
+		// grab nanoFramework.Runtime.Events.pe from the packages folder (it has to be there because the NF.TestApplication has just build)
+		// ************************************************************************* //
+		// just need to update the path on the package folder as the version changes //
+		// ************************************************************************* //
+		vec.push_back(L"-load");
+		vec.push_back(L"..\\packages\\nanoFramework.Runtime.Events.1.0.0-preview007\\lib\\nanoFramework.Runtime.Events.pe");
 
 		// grab Windows.Devices.Spi.pe from the packages folder (it has to be there because the NF.TestApplication has just build)
 		// ************************************************************************* //
 		// just need to update the path on the package folder as the version changes //
 		// ************************************************************************* //
 		vec.push_back(L"-load");
-		vec.push_back(L"..\\packages\\nanoFramework.Windows.Devices.Spi.1.0.0-preview003\\lib\\Windows.Devices.Spi.pe");
+		vec.push_back(L"..\\packages\\nanoFramework.Windows.Devices.Spi.1.0.0-preview006\\lib\\Windows.Devices.Spi.pe");
 
 
         //vec.push_back(L"-load");
