+++ targets/os/win32/netnf/TestApplication/NF.TestApplication.nfproj
@@ -27,13 +27,18 @@
       <Private>True</Private>
       <SpecificVersion>True</SpecificVersion>
     </Reference>
+    <Reference Include="nanoFramework.Runtime.Events, Version=1.0.0.0, Culture=neutral, PublicKeyToken=72027bbd7f714be2">
+      <HintPath>..\..\packages\nanoFramework.Runtime.Events.1.0.0-preview007\lib\nanoFramework.Runtime.Events.dll</HintPath>
+      <Private>True</Private>
+      <SpecificVersion>True</SpecificVersion>
+    </Reference>
     <Reference Include="Windows.Devices.Gpio, Version=1.0.0.0, Culture=neutral, PublicKeyToken=6df87852110734f7">
-      <HintPath>..\..\packages\nanoFramework.Windows.Devices.Gpio.1.0.0-preview006\lib\Windows.Devices.Gpio.dll</HintPath>
+      <HintPath>..\..\packages\nanoFramework.Windows.Devices.Gpio.1.0.0-preview010\lib\Windows.Devices.Gpio.dll</HintPath>
       <Private>True</Private>
       <SpecificVersion>True</SpecificVersion>
     </Reference>
     <Reference Include="Windows.Devices.Spi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=6df87852110734f7">
-      <HintPath>..\..\packages\nanoFramework.Windows.Devices.Spi.1.0.0-preview003\lib\Windows.Devices.Spi.dll</HintPath>
+      <HintPath>..\..\packages\nanoFramework.Windows.Devices.Spi.1.0.0-preview006\lib\Windows.Devices.Spi.dll</HintPath>
       <Private>True</Private>
       <SpecificVersion>True</SpecificVersion>
     </Reference>
