+++ targets/os/win32/netnf/TestApplication/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="nanoFramework.CoreLibrary" version="1.0.0-preview022" />
-  <package id="nanoFramework.Windows.Devices.Gpio" version="1.0.0-preview006" />
-  <package id="nanoFramework.Windows.Devices.Spi" version="1.0.0-preview003" />
+  <package id="nanoFramework.Runtime.Events" version="1.0.0-preview007" />
+  <package id="nanoFramework.Windows.Devices.Gpio" version="1.0.0-preview010" />
+  <package id="nanoFramework.Windows.Devices.Spi" version="1.0.0-preview006" />
 </packages>
